move aea.helpers.pypi module
move 'aea.helpers.pypi' module into aea.configurations, since it
imports form aea.configuration.base
